+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # V 0.0.0
+## 2015-03-30
+* added yolo-bear-users table in dynamodb + putNick.php to api + dropStaleNick.php to scripts
+
 ## 2015-02-28
 * moved documentation for running peerjs server from yolo-bear to yolo-bear-server
 * first implementation of api